+++ src/common/api.ts
@@ -29,12 +29,11 @@ import {
   UsageDataRow,
   EnvironmentVariable,
   CustomCommand,
-  Font,
   StartupMode,
-  Theme,
   FileEdit,
   ClearProjectData,
   ProjectStartedData,
+  CloudflareTunnelStatus,
 } from '@common/types';
 
 export interface ApplicationAPI {
@@ -42,11 +41,12 @@ export interface ApplicationAPI {
   openLogsDirectory: () => Promise<boolean>;
   loadSettings: () => Promise<SettingsData>;
   saveSettings: (settings: SettingsData) => Promise<SettingsData>;
-  saveTheme: (theme: Theme) => Promise<Theme>;
-  saveFont: (font: string) => Promise<Font>;
   isManageServerSupported: () => boolean;
   startServer: (username?: string, password?: string) => Promise<boolean>;
   stopServer: () => Promise<boolean>;
+  startCloudflareTunnel: () => Promise<boolean>;
+  stopCloudflareTunnel: () => Promise<void>;
+  getCloudflareTunnelStatus: () => Promise<CloudflareTunnelStatus>;
   startProject: (baseDir: string) => void;
   stopProject: (baseDir: string) => void;
   restartProject: (baseDir: string, startupMode?: StartupMode) => void;